Refactor legend hook and update imports to use useChartLegendItems
