+++ projects/js-packages/charts/src/components/legend/index.ts
@@ -1,5 +1,5 @@
 export { Legend } from './legend';
 export { BaseLegend } from './base-legend';
-export { useChartLegendData } from './use-chart-legend-data';
+export { useChartLegendItems } from './use-chart-legend-items';
 export type { LegendProps, BaseLegendProps } from './types';
-export type { ChartLegendOptions } from './use-chart-legend-data';
+export type { ChartLegendOptions } from './use-chart-legend-items';
